+++ implementations/node/packages/cli/src/workspace.ts
@@ -157,6 +157,7 @@ export class Workspace {
       },
     );
     tenant.assetService.setRuleModule(ruleModules);
+    tenant.costing.setRuleModule(ruleModules);
     return tenant;
   }